Renamed ValueConfig in ValueConfiguration
